refactor: move navigation service to core
