+++ plugins-structure/apps/politeia/src/utils/routeCleanup.js
@@ -1,7 +1,7 @@
 import ReactDOM from "react-dom";
 import { store } from "@politeiagui/core";
 import { router } from "@politeiagui/core/router";
-import { navigation } from "@politeiagui/common-ui/services";
+import { navigation } from "@politeiagui/core/globalServices";
 
 export function routeCleanup() {
   const location = router.getCurrentLocation();
