--- plugins-structure/packages/common-ui/src/services/index.js
@@ -1,13 +1,0 @@
-import {
-  popNavigation,
-  pushNavigation,
-  selectNavigationHistory,
-  selectNavigationHistoryLastItem,
-} from "./navigation";
-
-export const navigation = {
-  pop: popNavigation,
-  push: pushNavigation,
-  selectHistory: selectNavigationHistory,
-  selectHistoryLastItem: selectNavigationHistoryLastItem,
-};
