+++ plugins-structure/packages/core/src/globalServices/index.js
@@ -11,6 +11,26 @@ import {
   selectMessage,
   setMessage,
 } from "./message";
+import {
+  services as navigationServices,
+  popNavigation,
+  pushNavigation,
+  selectNavigationHistory,
+  selectNavigationHistoryLastItem,
+} from "./navigation";
+
+export const message = {
+  clear: clearMessage,
+  select: selectMessage,
+  set: setMessage,
+};
+
+export const navigation = {
+  pop: popNavigation,
+  push: pushNavigation,
+  selectHistory: selectNavigationHistory,
+  selectHistoryLastItem: selectNavigationHistoryLastItem,
+};
 
 export const progress = {
   end: endProgress,
@@ -19,10 +39,8 @@ export const progress = {
   update: updateProgress,
 };
 
-export const message = {
-  clear: clearMessage,
-  select: selectMessage,
-  set: setMessage,
-};
-
-export const services = [...progressServices, ...messageServices];
+export const services = [
+  ...progressServices,
+  ...messageServices,
+  ...navigationServices,
+];
